Relocated hook interfaces to their respective hook files.
